# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TalkABC 聊天交友平台后端 API 文档",
        "title": "TalkABC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/auth/login/oauth": {
            "post": {
                "description": "使用第三方平台凭证完成用户登录，支持Apple、Google、微信、支付宝、Email五种登录方式\n\n**支持的登录方式及所需参数：**\n- Apple登录：需提供 id_token 参数\n- Google登录：需提供 id_token 参数\n- 微信登录：需提供 code 参数\n- 支付宝登录：需提供 code 参数\n- 邮箱登录：需提供 email 参数\n\n**安全防护机制：**\n- 凭证验证：验证第三方平台返回的凭证有效性\n- 新用户自动注册：首次登录时自动创建用户账号\n- 账号绑定检测：已绑定的第三方账号直接登录\n- 状态校验：验证用户账号状态（正常/封禁/注销）\n- 操作审计：记录完整的登录日志，日志不可删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "第三方登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录方式：apple/google/wechat/alipay/email",
                        "name": "provider",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码（微信、支付宝登录使用）",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "ID令牌（Apple、Google登录使用）",
                        "name": "id_token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "邮箱地址（邮箱登录使用）",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回access_token、refresh_token、new_user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误或验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "登录失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/auth/oauth/bind": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将第三方平台账号绑定到当前登录用户，绑定后可使用该第三方平台直接登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "绑定第三方账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "绑定方式：apple/google/wechat/alipay/email",
                        "name": "provider",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码（微信、支付宝绑定使用）",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "ID令牌（Apple、Google绑定使用）",
                        "name": "id_token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "邮箱地址（邮箱绑定使用）",
                        "name": "email",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误或验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "绑定失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/oauth/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户已绑定的所有第三方平台账号列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取已绑定的第三方账号列表",
                "responses": {
                    "200": {
                        "description": "获取成功，返回已绑定的provider列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/oauth/unbind": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "解绑当前登录用户的第三方平台账号，解绑后将无法使用该第三方平台登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "解绑第三方账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "解绑方式：apple/google/wechat/alipay/email",
                        "name": "provider",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "解绑成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "解绑失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}